% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internals.R
\name{getlogdata}
\alias{getlogdata}
\title{Get log data}
\usage{
getlogdata(datum)
}
\arguments{
\item{datum}{Name of datum to get}
}
\value{
Value of that datum
}
\description{
Get log data
}
\details{
This handles internal data tracking only, not the file on disk.
}
\keyword{internal}

