% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internals.R
\name{getloginfo}
\alias{getloginfo}
\title{Get current log info}
\usage{
getloginfo()
}
\value{
A list with information about current (active) log
}
\description{
Get current log info
}
\details{
This handles internal data tracking only, not the file on disk.
}
\keyword{internal}

