% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/luxlogr.R
\docType{package}
\name{luxlogr}
\alias{luxlogr}
\alias{luxlogr-package}
\title{Lightweight logging for R}
\description{
Logging facilities for R scripts. Very lightweight, but supports message
prioritization, echoing to screen (or not), timestamps, multiple open
logs, etc.
}

