% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/luzlogr.R
\docType{package}
\name{luzlogr}
\alias{luzlogr}
\alias{luzlogr-package}
\title{Lightweight logging for R}
\description{
This package provides flexible but lightweight logging facilities for R scripts.
Supports priority levels for logs and messages, flagging messages,
capturing script output, switching logs, and logging to files or connections.
}

