% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internals.R
\name{newlog}
\alias{newlog}
\title{Create new log}
\usage{
newlog(logfile, loglevel, sink, description, closeit)
}
\arguments{
\item{logfile}{Name of log file (character or connection)}

\item{loglevel}{Minimum priority level (numeric)}

\item{sink}{Send all console output to logfile? (logical)}

\item{description}{Description (character)}

\item{closeit}{File should be closed when log closes? (logical)}
}
\description{
Create new log
}
\details{
This handles internal data tracking only, not the file on disk.
}
\keyword{internal}

