% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logging.R
\name{outputdir}
\alias{outputdir}
\title{Return output directory}
\usage{
outputdir(scriptname, scriptfolder = TRUE)
}
\arguments{
\item{scriptname}{Name of script (or output folder name)}

\item{scriptfolder}{Script-specific output folder? (logical, optional)}
}
\value{
Output directory
}
\description{
Return output directory
}
\details{
Return output directory (perhaps inside a script-specific folder)
If caller specifies `scriptfolder=FALSE`, return OUTPUT_DIR
If caller specifies `scriptfolder=TRUE` (default), return OUTPUT_DIR/SCRIPTNAME
}
\keyword{internal}

