% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logging.R
\name{closelog}
\alias{closelog}
\title{Close current logfile}
\usage{
closelog(sessionInfo = TRUE)
}
\arguments{
\item{sessionInfo}{Print \code{\link{sessionInfo}} output? (logical, optional)}
}
\value{
Invisible success (TRUE) or failure (FALSE)
}
\description{
Close current logfile
}
\details{
Close current logfile
}
\seealso{
\code{\link{openlog}} \code{\link{printlog}}
}

