% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/luzlogr.R
\docType{package}
\name{luzlogr}
\alias{luzlogr}
\alias{luzlogr-package}
\title{Lightweight logging for R}
\description{
Logging facilities for R scripts. Very lightweight, but supports message
prioritization, echoing to screen (or not), timestamps, multiple open
logs, etc.
}

