% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internals.R
\name{newlog}
\alias{newlog}
\title{Create new log}
\usage{
newlog(logfile, loglevel, sink, closeit)
}
\arguments{
\item{logfile}{Name of log file}

\item{loglevel}{Minimum priority level (numeric, optional)}

\item{sink}{Send all console output to logfile? (logical, optional)}

\item{closeit}{File should be closed when log closes?}
}
\description{
Create new log
}
\details{
This handles internal data tracking only, not the file on disk.
}
\keyword{internal}

