% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internals.R
\name{removelog}
\alias{removelog}
\title{Remove current log}
\usage{
removelog()
}
\value{
The just-removed log info
}
\description{
Remove current log
}
\details{
This handles internal data tracking only, not the file on disk.
}
\keyword{internal}

