% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{setlogdata}
\alias{setlogdata}
\title{Set log data}
\usage{
setlogdata(datum, value)
}
\arguments{
\item{datum}{Name of datum to set (character)}

\item{value}{Value}
}
\description{
Set log data
}
\details{
This handles internal data tracking only, not the file on disk.
}
\keyword{internal}
